# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Employee Management System API
  description: |
    RESTful API for managing employees, users, leave requests, and authentication.
    
    Features:
    - Employee CRUD operations with CSV import/export
    - User management with role-based access control
    - Leave request workflow with approval/rejection
    - Email verification and password reset
    - Rate limiting and caching
  version: 1.0.0
  contact:
    name: EMS Development Team
  license:
    name: MIT License
servers:
  - url: https://employee-management-system-production-2a1a.up.railway.app/api/v1
    description: Production server (Railway)
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /users:
    get:
      tags:
        - User Management
      summary: Get all users
      operationId: getAllUsers
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
            maximum: 20
        - name: sortBy
          in: query
          schema:
            type: string
            default: id
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserDto'
  /users/{id}:
    get:
      tags:
        - User Management
      summary: Get user by ID
      operationId: getUserById
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

    put:
      tags:
        - User Management
      summary: Update user
      operationId: updateUser
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: User updated successfully

    delete:
      tags:
        - User Management
      summary: Delete user
      operationId: deleteUser
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted successfully
  /users/create-with-temp-password:
    post:
      tags:
        - User Management
      summary: Create user with temporary password
      operationId: createUserWithTemporaryPassword
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User created successfully
  /users/{id}/roles:
    put:
      tags:
        - User Management
      summary: Update user roles
      operationId: updateUserRoles
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: User roles updated successfully

  # Leave Management
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session-based authentication
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        userName:
          type: string
          example: "john.doe"
        enabled:
          type: boolean
          example: true
        temporaryPassword:
          type: boolean
          example: false
        emailVerified:
          type: boolean
          example: true
        roles:
          type: array
          items:
            type: string
    PagedUserDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalElements:
          type: integer
        totalPages:
          type: integer
    CreateUserRequest:
      type: object
      properties:
        employeeId:
          type: integer
        username:
          type: string
        temporaryPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        authorities:
          type: array
          items:
            type: string
    RoleUpdateRequest:
      type: object
      properties:
        userId:
          type: integer
        roleNames:
          type: array
          items:
            type: string
          example: ["MANAGER", "EMPLOYEE"]
...